cleanup the test spec
