+++ test/Stubs.ts
@@ -1,0 +1,3 @@
+export interface FsStub {
+    writeFile: any;
+};
